requirements.txt
Flask
Flask-RESTful
openai
python-dotenv
flask-swagger-ui
pandas
scikit-learn
gunicorn
